chore: product group attr

+++ ec-project-api/Models/products/ProductGroup.cs
@@ -12,7 +12,7 @@ namespace ec_project_api.Models {
         public string? Name { get; set; }
 
         [Column("status_id")]
-        public int StatusId { get; set; }
+        public short StatusId { get; set; }
 
         [Required]
         [Column("created_at")]
